implement max timeout for partial received message/missing packets

+++ morse/morse.cpp
@@ -54,7 +54,7 @@ void prepareDevice() {
   }
 }
 
-void blinkRedLED(int ms) {
+void blinkRedLED(unsigned long ms) {
   digitalWrite(PIN_RED, HIGH);
   delay(ms);
   digitalWrite(PIN_RED, LOW);
